+++ tests/basics/tokenizer/tokenizer.cpp
@@ -27,7 +27,7 @@ Tokenizer::next_token(const std::basic_string_view<char> input,
   aux = [&](unsigned int fuel, int64_t index, std::basic_string_view<char> s)
       -> std::pair<std::optional<std::basic_string_view<char>>,
                    std::basic_string_view<char>> {
-    if ((s.length() == 0)) {
+    if ((s.length() == int64_t(0))) {
       return std::make_pair(std::nullopt, std::string_view(nullptr, 0));
     } else {
       if (fuel <= 0) {
@@ -40,34 +40,34 @@ Tokenizer::next_token(const std::basic_string_view<char> input,
         if (hard.contains(c)) {
           return std::make_pair(
               std::make_optional<std::basic_string_view<char>>(
-                  s.substr(0, index)),
-              s.substr(
-                  ((index + 1) & 0x7FFFFFFFFFFFFFFFLL),
-                  ((input.length() - ((index + 1) & 0x7FFFFFFFFFFFFFFFLL)) &
-                   0x7FFFFFFFFFFFFFFFLL)));
+                  s.substr(int64_t(0), index)),
+              s.substr(((index + int64_t(1)) & 0x7FFFFFFFFFFFFFFFLL),
+                       ((input.length() -
+                         ((index + int64_t(1)) & 0x7FFFFFFFFFFFFFFFLL)) &
+                        0x7FFFFFFFFFFFFFFFLL)));
         } else {
           if (soft.contains(c)) {
-            if ((index == 0)) {
-              return aux(
-                  fuel_, 0,
-                  s.substr(1, ((input.length() - 1) & 0x7FFFFFFFFFFFFFFFLL)));
+            if ((index == int64_t(0))) {
+              return aux(fuel_, int64_t(0),
+                         s.substr(int64_t(1), ((input.length() - int64_t(1)) &
+                                               0x7FFFFFFFFFFFFFFFLL)));
             } else {
               return std::make_pair(
                   std::make_optional<std::basic_string_view<char>>(
-                      s.substr(0, index)),
-                  s.substr(
-                      ((index + 1) & 0x7FFFFFFFFFFFFFFFLL),
-                      ((input.length() - ((index + 1) & 0x7FFFFFFFFFFFFFFFLL)) &
-                       0x7FFFFFFFFFFFFFFFLL)));
+                      s.substr(int64_t(0), index)),
+                  s.substr(((index + int64_t(1)) & 0x7FFFFFFFFFFFFFFFLL),
+                           ((input.length() -
+                             ((index + int64_t(1)) & 0x7FFFFFFFFFFFFFFFLL)) &
+                            0x7FFFFFFFFFFFFFFFLL)));
             }
           } else {
-            return aux(fuel_, ((index + 1) & 0x7FFFFFFFFFFFFFFFLL), s);
+            return aux(fuel_, ((index + int64_t(1)) & 0x7FFFFFFFFFFFFFFFLL), s);
           }
         }
       }
     }
   };
-  return aux(static_cast<unsigned int>(input.length()), 0, input);
+  return aux(static_cast<unsigned int>(input.length()), int64_t(0), input);
 }
 
 std::shared_ptr<List::list<std::basic_string_view<char>>>
