+++ tests/wip/int63_arith/int63_arith.cpp
@@ -13,8 +13,8 @@
 #include <variant>
 
 int64_t Int63Arith::i_abs(const int64_t x) {
-  if ((x < 0)) {
-    return ((0 - x) & 0x7FFFFFFFFFFFFFFFLL);
+  if ((x < int64_t(0))) {
+    return ((int64_t(0) - x) & 0x7FFFFFFFFFFFFFFFLL);
   } else {
     return x;
   }
