+++ tests/wip/string_match/string_match.cpp
@@ -12,4 +12,6 @@
 #include <utility>
 #include <variant>
 
-bool StringMatch::is_empty(const std::string s) { return (s.length() == 0); }
+bool StringMatch::is_empty(const std::string s) {
+  return (s.length() == int64_t(0));
+}
